clean up assignment page and add example with proper assignment for tscircutisch04
